Re-organized file-structure for project.
Moved private/internal headers into source dir and out of publically
exported directories. Moved src into a `pembroke` sub-folder to match
the include directory and provide consistent include declarations to
private headers.
