+++ pembroke/CMakeLists.txt
@@ -8,13 +8,13 @@ conan_basic_setup()
 ## Pembroke Library Definition
 ## ---
 add_library(pembroke
-    src/buffer.cpp
-    src/event/delayed.cpp
-    src/event/timer.cpp
-    src/http/request.cpp
-    src/internal/logging.cpp
-    src/internal/util.cpp
-    src/reactor.cpp
+    src/pembroke/buffer.cpp
+    src/pembroke/event/delayed.cpp
+    src/pembroke/event/timer.cpp
+    src/pembroke/http/request.cpp
+    src/pembroke/internal/logging.cpp
+    src/pembroke/internal/util.cpp
+    src/pembroke/reactor.cpp
 )
 set_property(TARGET pembroke PROPERTY CXX_STANDARD 17)
 target_link_libraries(pembroke
@@ -24,6 +24,7 @@ target_link_libraries(pembroke
 target_include_directories(pembroke
     INTERFACE  ${CMAKE_CURRENT_SOURCE_DIR}/include
     PUBLIC     ${CMAKE_CURRENT_SOURCE_DIR}/include
+    PRIVATE    ${CMAKE_CURRENT_SOURCE_DIR}/src
 )
 
 
@@ -32,21 +33,25 @@ target_include_directories(pembroke
 ## ---
 find_package(Catch2 REQUIRED)
 add_executable(tests
-    src/main_test.cpp
+    src/pembroke/main_test.cpp
 
-    src/buffer_test.cpp
-    src/event/delayed_test.cpp
-    src/event/timer_test.cpp
-    src/internal/logging_test.cpp
-    src/internal/util_test.cpp
-    src/reactor_test.cpp
+    src/pembroke/buffer_test.cpp
+    src/pembroke/event/delayed_test.cpp
+    src/pembroke/event/timer_test.cpp
+    src/pembroke/internal/logging_test.cpp
+    src/pembroke/internal/util_test.cpp
+    src/pembroke/reactor_test.cpp
 
-    src/internal/test_common.cpp
+    src/pembroke/internal/test_common.cpp
 )
 set_property(TARGET tests PROPERTY CXX_STANDARD 17)
 target_link_libraries(tests 
     Catch2::Catch2
     pembroke
+)
+target_include_directories(tests
+    # allow tests to use internal headers
+    PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/src
 )
 include(Catch)
 catch_discover_tests(tests)
