+++ pembroke/src/pembroke/internal/logging.cpp
