+++ pembroke/src/pembroke/internal/util.cpp
